GUI3/requirements.txt
watchdog-0.10.3
pyqt5-5.15.0
